[Security Solution] unwire dead sourcerer redux slice from the store
The sourcerer redux slice is no longer read by any runtime code. Decouple
the common store, the store factory and the global_state test mock from it:
remove the slice from the State type and combineReducers, drop the preloaded
sourcerer state (and its skip:true createDefaultDataView seeding), and remove
the sourcerer state sanitizer. Test fixtures that referenced
mockGlobalState.sourcerer are updated to source their values elsewhere.

The now-orphaned slice files (actions/reducer/helpers/barrel) are deleted in
a follow-up commit; model.ts is kept in place as a shared-types file.
