+++ x-pack/solutions/security/plugins/security_solution/public/common/store/model.ts
@@ -8,5 +8,4 @@
 export { appModel } from './app';
 export { dragAndDropModel } from './drag_and_drop';
 export { inputsModel } from './inputs';
-export { sourcererModel } from '../../sourcerer/store';
 export type * from './types';
